separated data time functions into helpers file
